requirements/base.txt
pytz==2022.7  # https://github.com/stub42/pytz
Pillow  # https://github.com/python-pillow/Pillow        
whitenoise==6.2.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==4.0.8  # pyup: < 4.1  # https://www.djangoproject.com/
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-filter==23.5 # https://github.com/carltongibson/django-filter
django-modeltranslation==0.18.11 # https://github.com/Buren/django-modeltranslation

# PDF generator
reportlab==4.0.4
xhtml2pdf==0.2.15

# Customize django admin
django-jet-reboot==1.3.5

# Environment variable
python-decouple==3.8

# Payments
stripe==5.5.0
gopay==2.0.1
legacy-cgi==2.6.3
setuptools>=40.0.0,<76.0.0